Implement query methods through adapter
