+++ lib/jsonb_accessor.rb
@@ -6,10 +6,9 @@ require "active_record/connection_adapters/postgresql_adapter"
 
 require "jsonb_accessor/version"
 require "jsonb_accessor/helpers"
+require "jsonb_accessor/adapters"
 require "jsonb_accessor/macro"
 require "jsonb_accessor/query_helper"
-require "jsonb_accessor/query_builder"
-require "jsonb_accessor/attribute_query_methods"
 
 module JsonbAccessor
   extend ActiveSupport::Concern
