+++ lib/jsonb_accessor/adapters.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+require_relative "adapters/postgresql_adapter"
+require_relative "adapters/sqlite_adapter"
+
+module JsonbAccessor::Adapters
+end
