+++ lib/jsonb_accessor/adapters/sqlite_adapter.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module JsonbAccessor::Adapters::BaseAdapter
+  module_function
+
+  ATTRIBUTE_TYPE = :json
+
+  def apply(_model, _attribute); end
+end
